refactor(mimir): split ResearchService into domain service modules
Pure structural move: the 1546-line ResearchService god class is split
into five domain modules under src/services/ (paper, library, experiment,
server, wiki-admin) plus a shared common.ts (ServiceState + result
helpers). service.ts becomes a thin facade keeping all Remote signatures,
the 'research' namespace, ResearchServiceConfig, and the Context
augmentation intact; every method body now forwards to a domain function.

Wire contract is untouched: no method name, parameter, return type, or
failure code changes. Mutable instance state (compileStatus map, jobSeq
counter) rides a per-instance ServiceState object passed explicitly, so
new ResearchService instances stay isolated (no module-level state).

Rebased onto upstream main (ceb13a8, v0.4.0): the two newly added
Remotes (convertFigure, saveFigure) are forwarded to the experiment
domain module, and the server module picks up writeBackExperiment /
jobSummaryOf / appendExperimentLog. svg conversion deps ride the same
explicit deps object.
