+++ packages/mimir/src/services/library.ts
@@ -1,0 +1,228 @@
+/**
+ * Library domain module: the literature workbench (papers table CRUD, arXiv
+ * search/import, and the workspace PDF fetch). Thin forwarding of the
+ * `library.*` Remote namespace lives in `service.ts`.
+ * @module dsh-mimir/src/services/library
+ */
+
+import { randomUUID } from 'node:crypto'
+import { mkdir, rename, unlink, writeFile } from 'node:fs/promises'
+import { join } from 'node:path'
+import { fetchArxivPdf, fetchArxivSearch, paperPdfFileName } from '../tools/arxiv.ts'
+import type { ResearchWikiDomain } from '../store.ts'
+import type {
+  ArxivEntry,
+  PaperRecord,
+  ResearchFetchPaperPdfResult,
+  ResearchImportPaperResult,
+  ResearchPapersResult,
+  ResearchRemovePaperResult,
+  ResearchSearchArxivResult,
+  ResearchUpdatePaperResult,
+} from '../types.ts'
+import { rejected, success } from './common.ts'
+
+/** Everything the Library domain functions need from the service scope. */
+export interface LibraryDeps {
+  readonly workspaceDir: string
+  readonly domain: ResearchWikiDomain
+}
+
+/** Timeout of one arXiv API request made on the panel's behalf. */
+const ARXIV_FETCH_TIMEOUT_MS = 15_000
+/** Timeout of one panel-driven arXiv PDF download. */
+const ARXIV_PDF_FETCH_TIMEOUT_MS = 60_000
+/** Workspace-relative directory the fetched paper PDFs land in. */
+const PAPER_PDF_DIR = 'papers'
+/** Default result cap of one panel-driven arXiv search. */
+const ARXIV_SEARCH_DEFAULT_MAX_RESULTS = 10
+/** Hard result cap of one panel-driven arXiv search. */
+const ARXIV_SEARCH_MAX_RESULTS = 50
+
+/** Atomically replace one binary file through a unique same-directory sibling. */
+async function writeBytesAtomic(filePath: string, bytes: Uint8Array): Promise<void> {
+  const tempPath = `${filePath}.${randomUUID()}.tmp`
+  try {
+    await writeFile(tempPath, bytes, { mode: 0o666 })
+    await rename(tempPath, filePath)
+  } catch (error) {
+    await unlink(tempPath).catch(() => {})
+    throw error
+  }
+}
+
+/**
+ * List every remembered paper, most recently added first.
+ * @param deps - open wiki domain.
+ * @returns the literature cards for the panel's papers view.
+ */
+export function listPapers(deps: LibraryDeps): Promise<ResearchPapersResult> {
+  const papers = [...deps.domain.table('papers').entries()]
+    .map(([, record]) => record)
+    .sort((left, right) => right.addedAt.localeCompare(left.addedAt))
+  return Promise.resolve(success({ papers: Object.freeze(papers) }))
+}
+
+/**
+ * Search arXiv on the panel's behalf. The query must be non-empty
+ * (`invalid-input` otherwise); the request carries a hard 15s timeout and
+ * transport/HTTP failures settle as `operation-failed` with the underlying
+ * message.
+ * @param deps - retained for signature uniformity (future arXiv knobs); the
+ * call itself is pure network.
+ * @param request - the free-text query and an optional result cap
+ * (default 10, hard cap 50).
+ * @returns the parsed entries, newest API order preserved.
+ */
+export async function searchArxiv(
+  deps: LibraryDeps,
+  request: { query: string; maxResults?: number },
+): Promise<ResearchSearchArxivResult> {
+  const query = request.query.trim()
+  if (query === '') return rejected({ code: 'invalid-input', message: 'query must be non-empty' })
+  const maxResults = request.maxResults ?? ARXIV_SEARCH_DEFAULT_MAX_RESULTS
+  if (!Number.isSafeInteger(maxResults) || maxResults < 1 || maxResults > ARXIV_SEARCH_MAX_RESULTS) {
+    return rejected({ code: 'invalid-input', message: `maxResults must be an integer between 1 and ${ARXIV_SEARCH_MAX_RESULTS}` })
+  }
+  try {
+    const results = await fetchArxivSearch(query, maxResults, AbortSignal.timeout(ARXIV_FETCH_TIMEOUT_MS))
+    return success({ results: Object.freeze(results) })
+  } catch (error) {
+    return rejected({
+      code: 'operation-failed',
+      message: error instanceof Error ? error.message : 'arXiv search failed',
+    })
+  }
+}
+
+/**
+ * Remember one arXiv entry in the wiki's papers table. The write is an
+ * idempotent upsert keyed by the bare arXiv id: a re-import refreshes the
+ * metadata but preserves the existing record's notes and first-write
+ * timestamp.
+ * @param deps - open wiki domain.
+ * @param request - the parsed entry (an empty id or title is `invalid-input`).
+ * @returns whether the paper was newly imported (false on a refresh).
+ */
+export async function importPaper(
+  deps: LibraryDeps,
+  request: { entry: ArxivEntry },
+): Promise<ResearchImportPaperResult> {
+  const entry = request.entry
+  const arxivId = entry.id.trim()
+  if (arxivId === '' || entry.title.trim() === '') {
+    return rejected({ code: 'invalid-input', message: 'entry id and title must be non-empty' })
+  }
+  const table = deps.domain.table('papers')
+  const existing = table.get(arxivId)
+  const record: PaperRecord = {
+    arxivId,
+    title: entry.title,
+    authors: [...entry.authors],
+    summary: entry.summary,
+    url: entry.url === '' ? `https://arxiv.org/abs/${arxivId}` : entry.url,
+    notes: existing?.notes ?? '',
+    // A re-import refreshes the arXiv metadata but never wipes the
+    // workbench-curated organization fields.
+    tags: [...(existing?.tags ?? [])],
+    projectIds: [...(existing?.projectIds ?? [])],
+    addedAt: existing?.addedAt ?? new Date().toISOString(),
+  }
+  await table.put(arxivId, record)
+  return success({ imported: existing === undefined })
+}
+
+/**
+ * Remove one remembered paper; an unknown arXiv id is `paper-not-found`.
+ * @param deps - open wiki domain.
+ * @param request - the bare arXiv id.
+ * @returns the removed id.
+ */
+export async function removePaper(
+  deps: LibraryDeps,
+  request: { arxivId: string },
+): Promise<ResearchRemovePaperResult> {
+  const table = deps.domain.table('papers')
+  if (table.get(request.arxivId) === undefined) {
+    return rejected({ code: 'paper-not-found' })
+  }
+  await table.delete(request.arxivId)
+  return success({ arxivId: request.arxivId })
+}
+
+/**
+ * Partially update one remembered paper's organization fields: only the
+ * present fields (`tags`, `projectIds`, `notes`) change. Tags are trimmed,
+ * emptied out, and deduped; every linked project id must exist in the wiki
+ * (`invalid-input` otherwise). An unknown arXiv id is `paper-not-found`.
+ * @param deps - open wiki domain.
+ * @param request - the bare arXiv id plus the fields to replace.
+ * @returns the stored record after the update.
+ */
+export async function updatePaper(
+  deps: LibraryDeps,
+  request: {
+    arxivId: string
+    tags?: string[] | undefined
+    projectIds?: string[] | undefined
+    notes?: string | undefined
+  },
+): Promise<ResearchUpdatePaperResult> {
+  const table = deps.domain.table('papers')
+  const existing = table.get(request.arxivId)
+  if (existing === undefined) return rejected({ code: 'paper-not-found' })
+  if (request.projectIds !== undefined) {
+    for (const projectId of request.projectIds) {
+      if (deps.domain.table('projects').get(projectId) === undefined) {
+        return rejected({ code: 'invalid-input', message: `unknown project: ${projectId}` })
+      }
+    }
+  }
+  const next: PaperRecord = {
+    ...existing,
+    tags: request.tags === undefined
+      ? existing.tags
+      : [...new Set(request.tags.map(tag => tag.trim()).filter(tag => tag !== ''))],
+    projectIds: request.projectIds ?? existing.projectIds,
+    notes: request.notes ?? existing.notes,
+  }
+  await table.put(request.arxivId, next)
+  return success({ paper: next })
+}
+
+/**
+ * Download one remembered paper's arXiv PDF into the workspace and link it
+ * on the record: the bytes land at `papers/<arxiv id>.pdf` under the
+ * workspace root (same-name overwrite on a refetch, so a new arXiv version
+ * replaces the stale copy) and the record's `pdfPath` points at it. The
+ * panel reads the file back through the `/research/paper-pdf/<id>` route.
+ * An unknown arXiv id is `paper-not-found`; transport/HTTP/oversize
+ * failures settle as `operation-failed` with the underlying message.
+ * @param deps - workspace root and open wiki domain.
+ * @param request - the bare arXiv id.
+ * @returns the stored record after the update.
+ */
+export async function fetchPaperPdf(
+  deps: LibraryDeps,
+  request: { arxivId: string },
+): Promise<ResearchFetchPaperPdfResult> {
+  const table = deps.domain.table('papers')
+  const existing = table.get(request.arxivId)
+  if (existing === undefined) return rejected({ code: 'paper-not-found' })
+  let bytes: Uint8Array
+  try {
+    bytes = await fetchArxivPdf(request.arxivId, AbortSignal.timeout(ARXIV_PDF_FETCH_TIMEOUT_MS))
+  } catch (error) {
+    return rejected({
+      code: 'operation-failed',
+      message: error instanceof Error ? error.message : 'arXiv PDF download failed',
+    })
+  }
+  const relPath = `${PAPER_PDF_DIR}/${paperPdfFileName(request.arxivId)}`
+  const dir = join(deps.workspaceDir, PAPER_PDF_DIR)
+  await mkdir(dir, { recursive: true })
+  await writeBytesAtomic(join(dir, paperPdfFileName(request.arxivId)), bytes)
+  const next: PaperRecord = { ...existing, pdfPath: relPath }
+  await table.put(request.arxivId, next)
+  return success({ paper: next })
+}
